requirements.txt
opencv-python-headless
pytesseract
mysql-connector-python
pyautogui
selenium
pygetwindow
pytesseract
qrcode